Moved project to target name and added proper handling of targets (#514)
* Changed project--> target going along with cmake standards

* Removing unused backup and restore actions

* Addressing PR feedback

* Fixing merged in files

* Updated the type of the FileInfoUtils_InsertFileInfoIntoArray function

* Removed bad find_package

* Fixed more type conversion pains

* still fighting type conversions

+++ src/diagnostics_component/utils/file_info_utils/inc/file_info_utils.h
@@ -23,7 +23,7 @@ EXTERN_C_BEGIN
  */
 typedef struct tagFileInfo
 {
-    unsigned long fileSize; //!< size of the file in bytes
+    long long fileSize; //!< size of the file in bytes
     char* fileName; //!< the name of the file
     time_t lastWrite; //!< the last time the file was modified
 } FileInfo;
@@ -35,7 +35,7 @@ bool FileInfoUtils_InsertFileInfoIntoArray(
     FileInfo* sortedLogFiles,
     size_t sortedLogFileLength,
     const char* candidateFileName,
-    unsigned long sizeOfCandidateFile,
+    long long sizeOfCandidateFile,
     time_t candidateLastWrite);
 
 EXTERN_C_END
